requirements.txt
dotenv
gensim
librosa
matplotlib
networkx
node2vec
numpy
pillow
pinecone
sentence-transformers
torch
torchvision